# src/main/java/com/lucas/msla/dto/ConvertRequestDTO.java
package com.lucas.msla.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import lombok.*;

import java.math.BigDecimal;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ConvertRequestDTO {

    @NotBlank
    private String from;

    @NotBlank
    private String to;

    @NotNull
    @Positive
    private BigDecimal amount;
}